Refactor: split single-file factory tests into three separate files
